feat(nix): update flake dependencies and adjust configurations
This commit updates the flake.lock file with the latest versions of
dependencies. It also adjusts the configuration files to match the
changes in the updated dependencies. Some unused dependencies are
removed from the flake.nix file. The nix package is forced to the
latest version in the nix.nix file. The home-manager module is updated
to use the latest version of the home-manager flake. The neovim
configuration is updated to match the changes in the updated plugins.
The nix-index module is moved from the NixOS configuration to the
home-manager configuration.