refactor(nixos): restructure module configuration files
This commit restructures the NixOS module configuration files. The
default.nix files in the config subdirectories have been renamed and
moved up one level for clarity and ease of access.

Additionally, the default.nix file in the config directory has been
deleted, and its contents have been moved to the default.nix file in the
nixos directory. This change simplifies the structure and makes the
configuration more straightforward.

This restructuring should make the configuration easier to navigate and
maintain.